refactor(web): promote RootLayout to top-level + fix delete-account redirect
Move RootLayout from src/components/layout/root-layout.tsx to
src/root-layout.tsx. It's the only truly cross-cutting layout — the
domain layouts (chat, settings, logs, intelligence) stay colocated
with the code they depend on. Kept as a flat file rather than a
src/layouts/ directory since there's just one.

Also fix delete-account-section.tsx, which called navigate("/") after
logout — that path doesn't match any route in the tree and falls
through to NotFound. Redirect to routes.account.login instead, matching
the behavior of logout-page.tsx.

The Vite base change to /assistant/ in #31525 does NOT require a
React Router basename. Routes intentionally span /account, /logout,
and /assistant/* as absolute browser paths (see comment in routes.tsx).
Vite's base only affects asset URL emission, which #31538 already
handled via publicAsset().
