+++ apps/web/src/domains/chat/components/mobile-document-overlay.tsx
@@ -15,7 +15,7 @@ interface MobileDocumentOverlayProps {
  * surface referenced from chat.
  *
  * **Mounting constraint**: must render outside `RootLayout`'s inner
- * transformed wrapper (see `src/components/layout/root-layout.tsx`) so
+ * transformed wrapper (see `src/root-layout.tsx`) so
  * `position: fixed` anchors to the viewport's initial containing block
  * rather than the keyboard-following transform `RootLayout` applies when
  * the soft keyboard opens.
